backend/requirements.txt
# Core Framework
fastapi>=0.115.6
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.2.0

# Database & Storage
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt>=2.12.0
cryptography>=44.0.1
authlib>=1.3.1

# Utilities
python-dotenv==1.0.0
python-multipart>=0.0.18
httpx==0.26.0

# ML Models for Security
gliner>=0.2.0                    # GLiNER for PII detection
onnxruntime>=1.16.0              # ONNX runtime for optimized inference
transformers>=4.35.0             # Hugging Face transformers (for DeBERTa)
# torch>=2.0.0                   # Installed separately via CPU-only index in Dockerfile (saves 3GB)
optimum[onnxruntime]>=1.16.0     # ONNX optimization for transformers
sentencepiece>=0.1.99            # Tokenizer for DeBERTa
accelerate>=0.20.0               # Required for transformers model loading

# Optional: Microsoft Presidio for PII (advanced)
# presidio-analyzer>=2.2.0
# presidio-anonymizer>=2.2.0

# Basic monitoring (optional)
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
